feat: simple address signer
