+++ packages/simple-signer/src/lib.rs
@@ -1,6 +1,11 @@
 #[cfg(feature = "state-transitions")]
 pub mod signer;
 
+#[cfg(feature = "state-transitions")]
+pub mod simple_address_signer;
+
 pub mod single_key_signer;
 
+#[cfg(feature = "state-transitions")]
+pub use simple_address_signer::SimpleAddressSigner;
 pub use single_key_signer::SingleKeySigner;
